Enable multiple articles by moving to a Jekyll _posts setup
So far, we only had a single article on the site, which was just a hard-coded page. We want to be able to publish multiple articles, and have them show up on an overview page, which is linked to in the main menu.

This patch introduces both of these changes, by moving to a Jekyll `_posts` setup and adding an `articles/index.md` page.